Downloads/NihinConnect/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
Pillow
django-cors-headers
channels
